Deployment/requirements.txt
huggingface_hub>=0.19.0
datasets>=2.15.0
transformers>=4.35.0
fastapi>=0.104.1
pydantic>=2.4.2
uvicorn>=0.24.0
numpy>=1.26.0
pandas>=2.1.1
scikit-learn>=1.3.2
stable-baselines3>=2.1.0
gymnasium>=0.28.1
python-dotenv>=1.0.0
schedule>=1.2.0
